Big refactor

+++ app/src/main/java/com/orka/myfinances/lib/extensions/models/BasketItem.kt
@@ -2,6 +2,6 @@ package com.orka.myfinances.lib.extensions.models
 
 import com.orka.myfinances.data.models.basket.BasketItem
 
-fun List<BasketItem>.getPrice(): Int {
+fun List<BasketItem>.getPrice(): Long {
     return sumOf { it.product.salePrice * it.amount }
 }
